requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bleach==6.2.0
bootstrap4==0.1.0
cffi==2.0.0
channels==4.2.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.6.0
django-environ==0.12.0
django-summernote==0.8.20.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
environ==1.0
git-filter-repo==2.47.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.1
mysqlclient==2.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.2.0
PyYAML==6.0.3
redis==5.2.1
referencing==0.36.2
rpds-py==0.27.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
webencodings==0.5.1
whitenoise==6.10.0
zope.interface==8.0
